requirements.txt
numpy
opencv-python
pyserial
fcntl
msvcrt
tty
argparse
matplotlib




